+++ runtime/pangolin/src/lib.rs
@@ -153,11 +153,11 @@ frame_support::construct_runtime! {
 		XcmpQueue: cumulus_pallet_xcmp_queue = 32,
 		PolkadotXcm: pallet_xcm = 33,
 		CumulusXcm: cumulus_pallet_xcm = 34,
+		EthereumXcm: pallet_ethereum_xcm = 44,
 		DmpQueue: cumulus_pallet_dmp_queue = 35,
 
 		// EVM stuff.
 		Ethereum: pallet_ethereum = 36,
-		EthereumXcm: pallet_ethereum_xcm = 44,
 		EVM: pallet_evm = 37,
 		MessageTransact: darwinia_message_transact = 38,
